Build and start the API

Build and start the API

api(
  queue_id = NULL,
  workers = 2,
  results_dir = tempdir(),
  inputs_dir = NULL,
  log_level = "info",
  health_check_interval = 0
)

Arguments

queue_id

ID of an existing queue to connect to, creates a new one if NULL

workers

Number of workers to spawn

results_dir

The dir for results to be saved to

inputs_dir

THe directory where input files are stored

log_level

The "lgr" log level to use

health_check_interval

Interval in seconds, after which the next time the redis connection is used the connection will be reset. 0 for no reconnection. This is used on cloud services where the TCP connections have an idle timeout and must be manually kept alive or restarted by the client.

Value

Running API