Get debug output from naomi

Get debug output from naomi

download_debug(
  id,
  server = NULL,
  dest = tempfile(),
  github_token = Sys.getenv("GITHUB_DEBUG_DOWNLOAD_TOKEN"),
  verbose = TRUE
)

Arguments

id

The model run id to download. This will be printed below the error message.

server

The url of the server. The default is to use the production naomi/hint/hintr instance. It is not possible to use the staging instance. You can change this if running locally.

dest

The destination for the downloaded data. The actual data will be unpacked into a directory corresponding to the run id within this, so it is safe to use a common directory.

verbose

Add a progress bar

Details

This uses GitHub authentication to pull the debug file. You need to

Make sure you are a member of the naomi-debug GitHub team

Generate a GitHub personal access token. Create a new fine-grained token. Set the resource owner to "hivtools" organisation. Under "Permissions" add "Members" read-only permission. And generate the token.

Pass the token to this function, fill in the interactive prompt or set as an environment variable "GITHUB_DEBUG_DOWNLOAD_TOKEN".